fix: i64 accumulator + repair pattern + stricter CI tests
Three related fixes to address detection failures from #8879:

1. i64 accumulator (CRITICAL bug discovered):

   distance_i16 retornava i32. Max dist para 14 dims i16 é 14*65534^2 ~6e10

   que OVERFLOWA i32 (max 2.1e9). Distâncias grandes viravam negativas

   → comparações erradas → vizinhos errados → FN.

   Mesmo o AVX2 com mask ainda podia produzir valores intermediários

   que somados dariam overflow i32. Promovido para i64 em todos os níveis:

     - distance_i16 returns i64

     - AVX2 usa _mm256_cvtepi32_epi64 para extender ANTES de somar

     - ds.distance, search_phase, IVF search todos i64

2. Repair pattern (inspirado em vinicius-piassa):

   Quando fraud_count == 0 (todos legit) ou fraud_count == k (todos fraud),

   a decisão binária (approved = fraud_score < 0.6) é TRAVADA: nenhum vizinho

   adicional pode mudá-la (no máximo move 1 slot, fica do mesmo lado do

   threshold de 3-of-5). Retorna imediatamente sem escanear mais células.

   Caso ambíguo (1..k-1): expande nprobe de 1 para 8 (IVF_NPROBE_REPAIR).

   Esperado: -50% p99 no caminho comum.

3. CI tests mais rigorosos (não vai submeter sem isso):

   - Adicionado teste caso claramente LEGÍTIMO (deve approved:true)

   - Adicionado teste caso claramente FRAUDE (deve approved:false)

   - Warnings se detectar errado (mas não falha para deixar gerar score)

+++ api/src/dataset.rs
@@ -160,7 +160,7 @@ impl Dataset {
         &self,
         query: &[i16; DIMS],
         idx: usize,
-    ) -> i32 {
+    ) -> i64 {
         distance_impl(query, idx, self.count, &self.dims)
     }
 
@@ -170,11 +170,11 @@ impl Dataset {
         query: &[i16; DIMS],
         k: usize,
     ) -> (Vec<usize>, usize) {
-        let mut best: Vec<(i32, usize)> = Vec::with_capacity(k);
-        
+        let mut best: Vec<(i64, usize)> = Vec::with_capacity(k);
+
         for i in 0..self.count {
             let dist = self.distance(query, i);
-            
+
             if best.len() < k {
                 best.push((dist, i));
                 let mut j = best.len() - 1;
@@ -191,16 +191,17 @@ impl Dataset {
                 }
             }
         }
-        
+
         let fraud_count = best.iter().filter(|(_, idx)| self.labels[*idx] != 0).count();
         let indices = best.into_iter().map(|(_, idx)| idx).collect();
         (indices, fraud_count)
     }
 }
 
-/// L2 distance implementation with AVX2 or scalar fallback
+/// L2 distance implementation with AVX2 or scalar fallback.
+/// Returns i64 to avoid overflow (max: 14 * 65534^2 ≈ 6e10).
 #[inline(always)]
-fn distance_impl(query: &[i16; DIMS], idx: usize, count: usize, dims: &[i16]) -> i32 {
+fn distance_impl(query: &[i16; DIMS], idx: usize, count: usize, dims: &[i16]) -> i64 {
     #[cfg(has_c_avx2)]
     {
         let mut qa = [0i16; 16];
@@ -213,9 +214,9 @@ fn distance_impl(query: &[i16; DIMS], idx: usize, count: usize, dims: &[i16]) ->
     }
     #[cfg(not(has_c_avx2))]
     {
-        let mut sum: i32 = 0;
+        let mut sum: i64 = 0;
         for d in 0..DIMS {
-            let diff = query[d] as i32 - dims[d * count + idx] as i32;
+            let diff = query[d] as i64 - dims[d * count + idx] as i64;
             sum += diff * diff;
         }
         sum
